move feature toggle checker to base
