+++ backend/dataall/core/environment/api/resolvers.py
@@ -22,7 +22,7 @@ from dataall.core.stacks.db.stack_repositories import Stack
 from dataall.core.vpc.db.vpc_repositories import Vpc
 from dataall.base.db import exceptions
 from dataall.core.permissions import permissions
-from dataall.core.feature_toggle_checker import is_feature_enabled
+from dataall.base.feature_toggle_checker import is_feature_enabled
 from dataall.base.utils.naming_convention import (
     NamingConventionService,
     NamingConventionPattern,
