+++ backend/dataall/modules/datasets/api/storage_location/resolvers.py
@@ -1,7 +1,7 @@
 from dataall.base.api.context import Context
 from dataall.modules.catalog.db.glossary_repositories import GlossaryRepository
 from dataall.base.db.exceptions import RequiredParameter
-from dataall.core.feature_toggle_checker import is_feature_enabled
+from dataall.base.feature_toggle_checker import is_feature_enabled
 from dataall.modules.datasets.services.dataset_location_service import DatasetLocationService
 from dataall.modules.datasets_base.db.dataset_models import DatasetStorageLocation, Dataset
